Renamed `kripodb similarities serve` to `kripodb serve`, as it now also serves the fragments
